Added a stable isotope table / table of elements implementation.
This implements a static table of contents that can be edited dynamically.
The initial contents are based on the contents of the ISOTOPE.DAT file
commonly used in the MS community and the table only contains entries
for the stable isotopes for every element.

Technically, the implementation delivers proxy objects that reference the
entry in the table of elements, and the whole workings are wrapped behind
a pImpl idiom. Slight overkill, but really convenient to use now.
Changes to the element entries will cause side effects (true to the
reference character of the proxies); this can be made more explicit if one
finds the time to split the Element class into a pair of Element and
MutableElement.  In that case, only the latter would provide setter methods.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>